server/requirements.txt
Flask==2.3.2
pyserial==3.5